SPLIT: OpenWeather request
This commit goes with OpenWeather request structures in separated file.
